consumer error handling improve; invalidate cache functionality improve: invalidate cache modes; assets service fixes

+++ src/lib/services/assets/repo/mod.rs
@@ -22,7 +22,7 @@ pub struct FindParams {
     pub smart: Option<bool>,
     pub verification_status_in: Option<Vec<VerificationStatus>>,
     pub asset_label_in: Option<Vec<AssetLabel>>,
-    pub limit: i64,
+    pub limit: u32,
     pub after: Option<String>,
 }
